metadataQueue in main-extract-async.ts
